move to legacy: weave_types
